+++ src/index.ts
@@ -25,10 +25,7 @@ export default class extends Controller {
   copy (event: Event): void {
     event.preventDefault()
 
-    this.sourceTarget.select()
-    document.execCommand('copy')
-
-    this.copied()
+    navigator.clipboard.writeText(this.sourceTarget.value).then(() => this.copied())
   }
 
   copied (): void {